cloud/cloud_function/requirements.txt
paho-mqtt==2.1.0
google-cloud-pubsub==2.23.0
functions-framework==3.5.0
